+++ src/main/java/com/zyc/zdh/controller/ZdhUnstructureController.java
@@ -50,19 +50,17 @@ public class ZdhUnstructureController extends BaseController{
     public Logger logger= LoggerFactory.getLogger(this.getClass());
 
     @Autowired
-    ZdhNginxMapper zdhNginxMapper;
+    private ZdhNginxMapper zdhNginxMapper;
     @Autowired
-    EtlTaskUpdateLogsMapper etlTaskUpdateLogsMapper;
+    private EtlTaskUpdateLogsMapper etlTaskUpdateLogsMapper;
     @Autowired
-    JarTaskMapper jarTaskMapper;
+    private JarFileMapper jarFileMapper;
     @Autowired
-    JarFileMapper jarFileMapper;
+    private EtlTaskUnstructureMapper etlTaskUnstructureMapper;
     @Autowired
-    EtlTaskUnstructureMapper etlTaskUnstructureMapper;
+    private DataSourcesMapper dataSourcesMapper;
     @Autowired
-    DataSourcesMapper dataSourcesMapper;
-    @Autowired
-    EtlTaskUnstructureLogMapper etlTaskUnstructureLogMapper;
+    private EtlTaskUnstructureLogMapper etlTaskUnstructureLogMapper;
     @Autowired
     private ZdhPermissionService zdhPermissionService;
 
@@ -195,6 +193,8 @@ public class ZdhUnstructureController extends BaseController{
             etlTaskUnstructureInfo.setIs_delete(Const.NOT_DELETE);
             debugInfo(etlTaskUnstructureInfo);
 
+            checkPermissionByProductAndDimGroup(zdhPermissionService, etlTaskUnstructureInfo.getProduct_code(), etlTaskUnstructureInfo.getDim_group());
+
             etlTaskUnstructureMapper.insertSelective(etlTaskUnstructureInfo);
 
 
@@ -236,12 +236,17 @@ public class ZdhUnstructureController extends BaseController{
             etlTaskUnstructureInfo.setIs_delete(Const.NOT_DELETE);
             debugInfo(etlTaskUnstructureInfo);
             String id=etlTaskUnstructureInfo.getId();
+
+            EtlTaskUnstructureInfo oldEtlTaskUnstructureInfo = etlTaskUnstructureMapper.selectByPrimaryKey(etlTaskUnstructureInfo.getId());
+
+            checkPermissionByProductAndDimGroup(zdhPermissionService, etlTaskUnstructureInfo.getProduct_code(), etlTaskUnstructureInfo.getDim_group());
+            checkPermissionByProductAndDimGroup(zdhPermissionService, oldEtlTaskUnstructureInfo.getProduct_code(), oldEtlTaskUnstructureInfo.getDim_group());
+
             etlTaskUnstructureMapper.updateByPrimaryKeySelective(etlTaskUnstructureInfo);
 
-            EtlTaskUnstructureInfo etui = etlTaskUnstructureMapper.selectByPrimaryKey(etlTaskUnstructureInfo.getId());
 
             if (etlTaskUnstructureInfo.getUpdate_context() != null && !etlTaskUnstructureInfo.getUpdate_context().equals("")
-                    && !etlTaskUnstructureInfo.getUpdate_context().equals(etui.getUpdate_context())) {
+                    && !etlTaskUnstructureInfo.getUpdate_context().equals(oldEtlTaskUnstructureInfo.getUpdate_context())) {
                 //插入更新日志表
                 EtlTaskUpdateLogs etlTaskUpdateLogs = new EtlTaskUpdateLogs();
                 etlTaskUpdateLogs.setId(etlTaskUnstructureInfo.getId());
@@ -517,6 +522,7 @@ public class ZdhUnstructureController extends BaseController{
     public ReturnInfo etl_task_unstructure_log_delete(String[] ids) {
 
         try{
+            checkPermissionByProductAndDimGroup(zdhPermissionService, etlTaskUnstructureLogMapper, etlTaskUnstructureLogMapper.getTable(), ids);
             etlTaskUnstructureLogMapper.deleteLogicByIds("etl_task_unstructure_log_info",ids, new Timestamp(new Date().getTime()));
             return ReturnInfo.build(RETURN_CODE.SUCCESS.getCode(),"删除成功", null);
         }catch (Exception e){
